fix(vendors): move VendorClaimError out of 'use server' file
Same lint / build constraint that surfaced on Phase 4 PR-B: Next.js
`'use server'` files can only export async functions; a class
export fails the Turbopack build. Move the error to a sibling
non-`'use server'` module and import it from the action module.
Tests import it from the dedicated errors module directly.
